chore: LSP: clean code & fix: game name lookup
